Use upgrade timeout timer to monitor sending upgrade responses.

+++ src/cio_websocket.c
@@ -43,7 +43,7 @@
 
 static const uint8_t WS_MASK_SET = 0x80;
 static const uint8_t WS_HEADER_FIN = 0x80;
-static const unsigned int WS_MID_FRAME_SIZE = 65535;
+static const size_t WS_MID_FRAME_SIZE = 65535;
 
 static const uint64_t close_timeout_ns = UINT64_C(10) * UINT64_C(1000) * UINT64_C(1000) * UINT64_C(1000);
